Initial commit of top-level script
NOAA-GFDL/MDTF-diagnostics#43

Revamp top-level script to use argparse's subparsers. Each subcommand
has a CLI built from a configuration file. The subcommands themselves,
and the functions they call, are configured in
subcommands_template.jsonc.

This depends on updates to cli.py, to be included in a future commit. It
also omits "mdtf info" functionality.